meltria: add meltria package to load dataset
